rename: entity type file name
