+++ src/types/entity.ts
