+++ src/types/index.ts
@@ -1,3 +1,3 @@
 export * from './utils';
-export * from './entities';
+export * from './entity';
 export * from './state';
